requirements.txt
"django>=1.8.5"
useful_inkleby==0.3
markdown
django-htmlmin
reportlab
pyPdf2
oscodepoint
django-bootstrap-form
unicodecsv